Polish `scheme_repl_server.cpp`
Move classes to their own files.
